Add validation to InstantiateTemplate and new tests
Updated InstantiateTemplateRequestHandler.cs:
- Added `using` directives for `System` and `System.ServiceModel`.
- Added validation checks in `Execute` method for `TemplateId`, `ObjectId`, and `ObjectType`.
- Throws `FaultException` if any validation fails.

Modified SharedTests.projitems:
- Added new test file `TestInstantiateTemplate.cs`.

Updated TestSendEmail.cs:
- Removed `TestInstantiateTemplateRequestReturnsInstantiateTemplateResponseWithEmail` test method.

Created TestInstantiateTemplate.cs:
- Added tests for `InstantiateTemplateRequest` validation and response.
