+++ tests/SharedTests/TestSendEmail.cs
@@ -1,6 +1,5 @@
 ﻿using DG.XrmFramework.BusinessDomain.ServiceContext;
 using Microsoft.Crm.Sdk.Messages;
-using System;
 using System.Linq;
 using System.ServiceModel;
 using Xunit;
@@ -356,22 +355,6 @@ namespace DG.XrmMockupTest
             };
 
             Assert.Throws<FaultException>(() => orgAdminUIService.Execute(sendEmailRequest));
-        }
-
-        [Fact]
-        public void TestInstantiateTemplateRequestReturnsInstantiateTemplateResponseWithEmail()
-        {
-            var templateRequest = new InstantiateTemplateRequest
-            {
-                TemplateId = Guid.NewGuid(),
-                ObjectId = Guid.NewGuid(),
-                ObjectType = "account"
-            };
-
-            var response = orgAdminUIService.Execute(templateRequest) as InstantiateTemplateResponse;
-
-            Assert.Single(response.EntityCollection.Entities);
-            Assert.Equal("email", response.EntityCollection.Entities[0].LogicalName);
         }
     }
 }
